Finally solve uobject issues?
At least under OSX, dlopening a library and a module that was linked
against that library does not do what we want.  For a start, the module
has a dependency toward the installed path of the library, so it does not
work with non-installed tree.  Second, if you install the library to
please the module, then at run time the library exists twice, and the
"global" variables exists twice two.  Needless to say that it does not
work.

So we need to stop link UObjects against libuobject.

But then, at least on OSX, we get error because some symbols of the
modules are not defined.  This is because of '-no-undefined'.  This flag
was installed to please Windows, but I believe that it makes sense only
for dynamic libraries, not for dlopen modules (Linux guys should stop
thinking it's the same).

Therefore get rid of both the linking against libuobject, and of
-no-undefined.  At this point, I have no idea how Windows will behave.

	* sdk/param.mk.in (LDFLAGS): Remove -no-undefined.
	($(OUTSHLIB): No longer link against libuobject.
	* src/tests/bin/uobject-check.as: Run the uobjects as dlopen
	modules to see if at least they answer to --version.