Try harder to use our own ltdl.h.
We have a problem on platforms where Boost headers and libltdl are
installed at the same place: to get access to the former might give us
access to the latter, i.e., using an installed ltdl.h instead of our
shipped one.

Since anyway there are already too many -I in here, it's simpler to remove
the -Ilibltdl, but the qualify the #include.

	* src/Makefile.am (urbi_launch_CPPFLAGS): Remove.
	* src/urbi-launch.cc: Include libltdl/ltdl.h.