Add fast versions of "is_a" and "type_check".
Most of the time, we call "is_a" and "type_check" with a fixed proto and
with dynamic atoms. We now provide faster templated versions of those two
subprograms so that the common case does not require the traversal of the
prototypes list, nor the caching of the already visited prototypes (which
is necessary because we support prototypes loops).

	* src/object/cxx-object.hh, src/object/cxx-object.hxx (type_check):
	New templated version.
	* src/object/object.hh, src/object/object.hxx (is_a): New templated
	version.