Allow skipping the build of "ast-dump".
"ast-dump" requires the Boost regex library. Adding a "--disable-programs"
to the kernel allows not to require any Boost library at build time, only
Boost headers are necessary. This eases the test of new compilers and
targets using Boost headers only.

The default has been set to "yes". This may not be a good idea because
it will also change the "sdk-remote" and the "libjpeg" default, but
let's try.

	* configure.ac (--enable-programs, BUILD_PROGRAMS): New.
	* src/Makefile.am: Honor "BUILD_PROGRAMS".