Handle modifiers on lvalue with an explicit target.
The problem:
Modifiers and trajectory generator are handled in the Binder.
The code calls ast::lvalue_wrap, which uses a parametric_ast with '=',
which generates an ast::Assign node, which is not handled by the Binder,
but by the Desugarer which was allready called.

The solution:
Write the ast obtained after expansion of the ast_lvalue_wrap parametric
ast by hand. Reinvoking rewrite was not possible, as processing is not
idenpotent.

	* src/binder/binder.cc (visit(Assignment)): Write the ast
	required to wrap a lvalue by hand if a trajectory generator on
	an lvalue with an explicit target is encountered.