Always compile the UObjects, and install them.
Before, uobjects were compiled during "make check", now they are compiled
at "make all", and checked at "make check".  We also install them.

In the process, I have disabled the tests of umake-remote.  I don't know
if it is worth working for it, as we want to have it actually wrap
urbi-launch and a dlopen-module.

We need to sort these uobjects and decide which one we want to keep.  And
maybe start adding useful uobjects (such as the "libc" one Matthieu often
refers to).

	* src/tests/uobjects: Move all the uobject directories to...
	* src/uobjects: here
	* src/uobjects/local.mk, src/uobjects/uobjects.mk: New.
	* src/Makefile.am, src/tests/bin/uobject-check.as,
	* src/tests/uobjects/local.mk: Adjust.