Fix static initialization problems.
UObject starter classes are instanciated at static-initialization time,
and were using a std::list that was also initialized that way.
This patch uses getter functions that dynamicaly create the list on
first call to avoid this problem. Those functions are not inlined to
avoid the underlying list being duplicated in all the generated shared
libraries.

	* include/urbi/ustarter.hh: Remvoe list static members.
	* include/urbi/ustarter.hxx: Remove inlined accessors, always
	* use the accessors .
	* src/libuco/uobject-hub-common.cc: Use the accessors.
	* src/libuco/ustarter.cc (object_list, objecthub_list):
	implement.