Clean Lobby/UConnection initialization.
object::State was meant to ensure extensibility of the data the Lobby
needs to keep, but it was never useful, so get rid of it.

The UConnection was initializing its lobby by hand, instead of delegating
the initialization to the lobby itself.  This requires to hand the
connection to the lobby, which is anyway what the lobby is meant for, so
there is no really increase of the impedance.

The ghostconnection was used as the connection associated to the Lobby
prototype.  This caused many problems because of circular dependencies in
the initialization of the kernel.  Since anyway it is forbidden to
actually use the connection of the Lobby prototype, we now pass connections
as pointers to the lobbies, and the Lobby prototype is handed 0.

	* src/object/state.hh, src/object/state.cc: Remove.
	* src/object/fwd.hh, src/object/local.mk: Adjust.
	* src/kernel/uconnection.cc (UConnection::UConnection): Pass this
	to the lobby_.
	Don't handle the lobby_'s slots, let it do it.
	(UConnection::~UConnection): Don't remove the lobby_'s "lobby"
	slot, its Object destructor reclaims all the slots anyway.
	* src/object/lobby.hh, src/object/lobby.cc (state_, value_type)
	(value_get): Replace with...
	(connection_, connection_type, connection_get): these.
	Adjust dependencies.
	(proto_make): Use 0 as connection.
	* src/object/system.cc, src/runner/interpreter-visit.hxx: Adjust
	to call connection_get.