Have urbi-console bounce to urbi-launch.
The point is to factor the machinery to locate the various DLLs.  Besides,
it provides a simplification for Windows.  On this architecture, it is
hard to run a binary linked against a shlib which is not in the directory
of the executable.  But we do not want to copy libuobject (which is
what urbi-console actually is) in bindir because we have two (one for
remote component, another for "plugins").  Since urbi-launch dlopens these
shlibs, using a qualified path, we do not have the same problems.

Therefore, we exchange shlib lookup by ld.so by a dlopen that we control.

	* src/urbi-console.cc: Rewrite to exec urbi-console.
	* src/Makefile.am: Do no link urbi-console with libuobject.