Update libport.
This should address the distcheck failures.  The problem is as follows.

In order to please windows, we set libdir=bindir on that architecture.
All our packages know that and do it, but packages such as jpeg should
not know that.  So when configuring it, we add libdir='${libdir}' to
DISTCHECK_CONFIGURE_FLAGS.  So in SDK-Remote's Makefile we have
DISTCHECK_CONFIGURE_FLAGS = libdir='${libdir}' that appears, and Make
performs the expansion of ${libdir} before calling configure, which is not
what we wanted.

	* libport: Update to get a better DISTCHECK_CONFIGURE_FLAGS.
	* configure.ac: Typo: pass libdir=$libdir, not $bindir.