Seriously optimize data passing between UObjects.
When an UObject writes data to an UVar, it is stored as an UValue, and
only converted to an urbiScript object on demand.
Furthermore, if enableBypass(true) is called on an UVar, UBinary, UImage
and USound data written to it is not copied, but passed as is to
callback functions (as long as they cast to reference types). As a
consequence, the data is only available while notify callbacks are
executed.

	* share/urbi/uobject.u: Add the UValue class, with bouncers for
	asString and asToplevelPrintable.
	* src/kernel/uobject.cc: .
	* src/kernel/uvalue-cast.cc: Handle object::UValue type.
	* src/object/uvar.cc,
	* src/object/uvar.hh (getter): New, like accessor() but
	optionaly returns the UValue and not its content.
	(accessor): If content of UVar is an UValue, return its content.
	Add the UValue class, encapsulating an rObject and an urbi::UValue.
	* sdk-remote: Update.