Partially handle remote UObjects destruction upon disconnection.
It's impossible to do it the k1-way, since in k1:

var foo = new myUOBject();
// Close the connection hosting myUOBject, and foo is gone

In k2, the UObject can be in slots everywhere, even in closed
variables.

This patch just removes the UObjects from 'uobjects''s slots, so that no new
instance can be created, and so that the remote UOBject can be inserted again.

At least what should be done in further patches is to throw nice errors
if an attempt to use a 'disconnected' object is made, and to safely
unregister from notifyChanges.

	* share/urbi/uobject.u: Add a 'lobby' slot in each remote UObject.
	Monitor Lobby.onDisconnect, remove all objects in uobjects with
	this lobby.