Fix a deadlock in notifyChange loop detection.
The loop-detection variable was a boolean local to the UVar.
So two Jobs could not be in notifyChange of the same UVar at
the same time. Worse, if a Job ended up frozen in one of
the notify callbacks, the notifyChange would end up thinking
it was always looping.

So use a per-job boolean, implemented as a set<Job*>.

	* src/object/uvar.cc (update_): Adapt.
	* src/object/uvar.hh: inChange_ is now a set<void*>.