Do not try to read port-file if we don't need it.
In -s mode, port-file is bounced to libuobject and will be written to.
In -p mode, port-file is read.
So do not read port-file in -s mode and do not abort if it does not exist,
becauese we do not care if it does not exist, since we will write to it.

	* src/urbi-launch.cc(main): Do that.