Set PACKET_SIZE to a huge value.
Before, the data to send was stored in UConnection's queue, and sent chunk by
chunk. The network layer was responsible for calling UConnection::continue_send
when it was ready to accept new data.
Now, libport::Socket is doing the buffering, and Connection is never calling
continue_send.
So as a quick patch, just set packet size to a huge value to force the
UConnection to send all the data at each write().

	* src/kernel/connection.hh: Set PACKET_SIZE to 1G.