uobject: Protect UrbiScript code injected by UObjects when binding.
We want the code declaring an UObject and the code declaring bindings from
the user ctor to be separated by '|'. But we want said code to be
terminated by ';', so that it gets executed. Before, main() or dispatch()
would send the ';', but now that we have code to bind a single UObject
in a context, me must send the ';' after calling the UObject ctor.

	* include/urbi/ucontext-impl.hh: New interface method instanciated.
	* include /urbi/ustarter.hxx (instanciate): Call instanciated()
	right after calling the UObject constructor.
	* src/libuobject/remoteucontextimpl.hh: Declare instanciated.
	* src/libuobject/uobject.cc (instanciated): Just send ';'.