parser: AstFactory.
Make a real AstFactory object instead of a set of freestanding functions.
The main point is to have virtual functions to implement different
factories.  Typically, on space starving devices, we should desugar on the
fly instead of being faithful.  Or maybe apply the desugaring immediately.

In a first step, we still use static member functions.

	* src/parser/ast-factory.hh, src/parser/ast-factory.cc
	(AstFactory): New genuine class.
	(ast_*): Rename as...
	(AstFactory::make_*): these.
	* src/ast/parametric-ast.cc,
	* src/binder/binder.cc,
	* src/parser/fwd.hh,
	* src/parser/local.mk,
	* src/parser/parser-impl.cc,
	* src/parser/parser-impl.hh,
	* src/parser/ugrammar.y,
	* src/rewrite/desugarer.cc,
	* src/rewrite/pattern-binder.cc,
	* src/rewrite/rescoper.cc:
	Adjust.