Move urbi::main into header.
A recent change made libuco be included in liburbi.  But libuco (meant for
both libuobjects) also includes the implementations of "urbi::main" that
bounce in the end to a urbi::main that the library should provide.
Liburbi does not provide one, and is not expected to.  So instead of
introducing yet another library with urbi::main, just put all the
implementation in the header.

	* src/libuco/urbi-main.cc: "Rename" as...
	* include/urbi/umain.hxx: this.
	* src/libuco/local.mk, include/local.mk: Adjust.
	* include/urbi/umain.hh: Remove URBI_SDK_API for functions that
	are no longer in libuco.