uobjects: Optimize data transfer to remote uobjects.
Context: UVars written to from the C++ side store an rUValue containing an
UValue. This UValue is only converted to an rObject on demand.

The idea is to override asPrintable on rUValue and use urbi::UValue::print
which is much faster that the generic asPrintable, only having to deal with
a fixed list of types.

	* share/urbi/uobject.u (UValue.asPrintable): Bounce to
	extractAsToplevelPrintable instead of extract.asToplevelPrintable.
	* src/object/uvar.cc,
	* src/object/uvar.hh (extractAsToplevelPrintable): Use UValue::print,
	unless cache exists and value is void.