tests: overhaul.
Instead of building a single binary which agregates all the liburbi
tests, make one executable per test, as we already do for libport for
instance.  This simplifies the infrastructure, makes it easier to
understand, and it should also finally make it possible to use
embedded INSTRUMENTFLAGS.  It was not before, since there was a single
binary: INSTRUMENTFLAGS applied to all the tests.

	* src/tests/bin/tests.hh, src/tests/bin/tests.cc: Overhaul so
	that each test is a standalone program.
	* src/tests/bin/liburbi-check.m4sh: Adjust.
	* src/tests/liburbi/ping.cc, src/tests/liburbi/removecallbacks.cc,
	* src/tests/liburbi/syncvalues.cc, src/tests/liburbi/values.cc: Adjust.
	* src/tests/liburbi/local.mk: Adjust.