doc: start using \refSlot.
perl -pi -e 's/\b([[:upper:]]\w+)\.([[:alpha:]]\w+)\b/\\refSlot{$1}{$2}/g'

There are many false positives this way (say CLIENT.INI and so forth).
Also, we must check that we are not inside a \lstinline.  And finally,
many targets are missing, because we currently don't put hypertarget
when we don't use urbiscriptapi's \item, which means, every single
definition that either use arguments (e.g., getSlot), or is not
written in pure letters (e.g., '+').

This will have to be addressed.

So in a first step, experiment here.

	* doc/tables/release-notes.tex: Use \refSlot.