uobjects: rename src/uobjects as src/urbi
This is so that the build tree and the install tree are alike.  To check
the build tree, we need to be able to set the URBI_UOBJECT_PATH so that it
find non-install uobjects.  Since they are loaded as "urbi/foo",
they need to be in an urbi/ directory.

	* src/uobjects: Rename as...
	* src/urbi: this.
	* src/Makefile.am, src/tests/uobjects/local.mk: Adjust.