connectionTag: do not apply to the Lobby itself.
Issue #1344.

We want not to tag the Shell itself.  As it derived from Interpreter,
which sets a tag on itself at construction time, pass a new argument to
neutralize this behavior.

Then, when an Interpreter is created from another, check if the latter is
actually a shell, in which case, push the Lobby's tag (which is the
connectionTag).

	* src/runner/interpreter.hh, src/runner/interpreter.cc
	(Interpreter): Do that.
	* src/runner/shell.cc: Ask our Interpreter part not to be tagged.
	* doc/specs/lobby.tex: Document this.