Shell: refactor
This is a reimplementation of d2f862e140de262d7fdc3f8ba57416cadc220cf4
with better properties.  Instead of fighting to avoid applying the
connectionTag to the top-level interpreter, make it immune to
StopException.

This becomes easy once we realize that too much of the handling of the
top-level loop is performed in Interpreter, instead of Shell, which is
precisely the top-level interpreter.

First, revert the changes of d2f862e140de262d7fdc3f8ba57416cadc220cf4.

	* tests: Update to the reverted master.
	* src/runner/interpreter.hh, src/runner/interpreter.cc
	(Interpreter::Interpreter): Remove the "tag" argument.
	Adjust callers.

Second, move the handling of the top-level Nary to Shell.  One benefit is
that the handling of exceptions is really performed at the top-level.

	* src/runner/interpreter.hh, src/runner/interpreter.cc
	(top_level_visit): Move to...
	* src/runner/shell.hh, src/runner/shell.cc (eval_print): here.
	Ignore the StopException.
	Yield after having call eval_print, to reproduce the behavior of
	the Nary visit.

One difference is that the top-level visit was performed on a Nary, and
the handling of exceptions etc. was performed when evaluating each member
of the Nary, not around the evaluation of the whole.  Yet, the Shell is
evaluating Nary's as a whole.  So instead of queueing a list of Nary's,
the Shell should concatenate the arriving Nary's to its command queue.
This way we have the same evaluation order.