One io_service to rule them all.
To optimize socket usage it is best to have them all share the same
io_service. But rObject Socket needs to poll from within the
scheduler since they require a valid runner. So we had to modify
the main loop to also poll from a Job.

This change required to reorganize initialization a bit, since our io_service
is now an rObject: No socket can be created before root_class_initialize is
called.

	* include/kernel/userver.hh: Do not store an io_service.
	* include/kernel/userver.hxx (get_io_service): Move...
	* src/kernel/userver.cc: ...here, return object::Socket's io service.
	(ctor): Move wake-up system initialization in initialize(). Explicitly
	call root_class_initialize.
	(initialize): Set System.pollLoop as scheduler idle job.
	* share/urbi/socket.u: Remove poll loop.
	* src/console-server.cc (main_loop): Remove poll code.
	(init): Store fast and interactive arguments in System, integrate
	stdin reading into the Socket system under !WIN32.
	* src/kernel/uobject.cc (write_and_unfreeze): Wake up the kernel.
	* src/object/object.cc (ctor): Do not call root_class_initialize.
	* src/object/system.cc (pollLoop, poll): New. Move poll code here.