UObject: Remove unwanted data transfer in remote mode.
Each time an UVar is initialized in remote mode, a notifyChange is
immediately installed. But now bound functions can take UVars as argument.
In that case temporary UVars are created, but the notifyChange they setup
was never cleared.
Ideally we want to switch to smart pointers for UVars, but there is no way
of doing that without a brand new API.

	* src/libuco/uvar-common.cc (~UVar): unnotify if UVar was a
	temporary.