UObject: more subtle loopback check code.
If a remote UObject writes to an UVar, we want no notifyChange on his lobby.
But if a remote UObject executes urbiscript that writes to an UVar, we want
notifyChange to trigger.
So remote UObject no longuer use = but use 'update', and the remote
notifyChange checks wether current lobby is target lobby, and wether
it was called through update().

	* share/urbi/uobject.u (update): New.
	(external.var callback): Do not send message only if called from
	update().