ast: ParameterizedAst.
ParametricAst is not reentrant, a well known issue.  It is used most of
the time as a static variable, with a variable part (the list of the
argument it is fed with).  This has occasionally caught us when we need to
desugar recursively some construct, since the inner and the outer use of
the ParametricAst share the same lists of arguments.

The fix is also well known: while the AST part should definitely remain
static, the argument part must not.  This is most probably what bit Csaba
Kertész.

	* src/ast/parametric-ast.hh, src/ast/parametric-ast.hxx,
	* src/ast/parametric-ast.cc (ParameterizedAst): New.
	The variable part of ParametricAst.