doxygen: revamp completely.
I think it is a failure to try to generate the Doxyfile right before
calling doxygen, and it's a failure to try to factor it too the
smallest bit.  So, first rename Doxyfile to be more Make-compliant
(i.e., use an extension alike what's going to be generated).  Then
factor the installation of html directories.  That's in build-aux, part
of libport.

	* libport: Update to get newest build-aux.
	* doc/local.mk: New.
	* Makefile.am: Use it.
	* configure.ac: Adjust to Doxyfile changes.
	* doc/Doxyfile.in: Rename as...
	* doc/sdk-remote.dox.in: this.
	Adjust to the specific case of SDK Remote.
	In particular avoid libjpeg and libport.