examples: give priority to Boost headers.
The builds of urbi-image and urbi-balltracking fail on Mac OS X on the
build farm.  The issue is that these programs use X11, so we pass the
-I to find them.  Unfortunately, that means -I/opt/local/include,
where there is also a newer Boost installed.  So we mixed Boost 1.38
libraries with Boost 1.44 headers.

Exchange the order of the CPP flags in order to find Boost 1.38 headers.

	* src/examples/local.mk: here.