getSlot: support "&a" instead of ".&a" for 'getSlot("a")'.
Now that "&a" is no longer possibly parsed as "{} & a", there is room for
this syntactic sugar.  There remain other issues yet: for instance
"return &a" or "throw &a" could be parsed as "{return} & a", or
"return (&a)".  Using the precedence directives of Bison, it is the latter
which is the parsed result.

	* src/parser/ugrammar.y (exp: ".&" id): Change into...
	(exp: "&" id): this.
	(nstmt, stmt.opt, exp.opt): Promote the expression argument
	over the statement construct.  I.e., accept "return&a" as
	"return (&a)".

	* doc/specs/code.tex, doc/specs/lang.tex,
	* share/urbi/call-message.u, share/urbi/exception.u,
	* share/urbi/lazy.u, share/urbi/lobby.u, share/urbi/pattern.u,
	* share/urbi/system.u, share/urbi/uobject.u,
	* share/urbi/updatehook-stack.u: Convert ".&a" to "&a".