Revert "Exception.Lookup: delay computation of misspellings"
This reverts commit 02e65ecd822f2847961d6b388ade83a0bb4bc598.

The point of this patch was to delay the computation of the closest
match until it is actually needed – after all, the exception might be
caught, in which case we don't care about that.

But there are two downsides.

First the computation may be performed several times if we issue the
error message several times.  This is unlikely, and it would be easy
to cache.

Second and foremost, if the exception is displayed after a while, the
slot names might have significantly changed, and the slot might now
exist.