UVar: hook atcond, reproduce the on-demand changed! mechanism.
UVar wants to control when changed! is emitted, and to disable
the default behavior. The way to do that is to override the default
changed slot. Do it better than before by reproducing the on-demand
aspect of the feature (Actualy also fix previous patch 0a9f027).

Also, UVar knows when it gets updated, so atcond should not trace inside
its getter function. Implement that.

	* src/object/uvar.hh: changed_ event, changed() slot.
	* src/object/uvar.cc: changed() creates the event and replace
	itself on first call. Emit changed only if present.
	(accessor): Disable at dependency tracking for the scope.
	* src/runner/runner.cc,
	* src/runner/runner.hh (dependency_log_set): New.