Matrix: not equality-comparable.
Boost does not provide comparison operators on Matrices and Vectors
(http://lists.boost.org/MailArchives/ublas/2005/05/0261.php), so we were
actually not using the features from EqualityComparable (otherwise we
have had missing-definition errors for operator==).

It turns out that MSVC is too greedy and wanted to instantiate the
functions of EqualityComparable, which ended in missing definitions
for operator== bw Matrix/Vector.

	* include/urbi/object/matrix.hh, include/urbi/object/vector.hh,
	* src/object/matrix.cc, src/object/vector.cc: No longer derive
	from EqualityComparable.