Objects: optional maps to nil, not void.
	* include/urbi/object/cxx-conversions.hxx: Do that.
	* include/urbi/object/dictionary.hh: Propose a conversion for
	Dictionaries.
	* src/object/system.cc: Use boost::optional instead of dealing
	with nil by hand.

This changes the return value of InputStream.get and others, in a
much more logical way.  Adjust to that, and document it.

	* doc/specs/input-stream.tex, doc/specs/stream.tex,
	* share/urbi/test-suite.u: Here.

But this is likely to break compatibility where people used to expect
void, so used isVoid instead of isNil.  Therefore, let nil.isVoid return
true, and warn.

	* share/urbi/nil.u, doc/tables/release-notes.tex:
	here.

InputStream: use Boost.Optional.

	* src/object/urbi/input-stream.cc,
	* src/object/urbi/input-stream.hh (getChar, getLine): Here.

Position: use Boost.Optional.

	* include/urbi/object/position.hh, src/object/position.cc
	(file_get): Here.
	* doc/specs/position.tex: Style changes.