parser: avoid cascading syntax errors on lexical errors.
Before, lexical errors were handled by the scanner, using the parser's
error reporting function.  As a result, something like "var 3x;"
resulted in '3x' causing a first syntax error, but was discarded by
the scanner.  So the parser received "VAR SEMI", so of course we had a
second error message for the unexpected semicolon:

  $ urbi -qe 'var 3x;shutdown;'
  [00000007:error] !!! 1.5-6: syntax error: invalid token: `3x'
  [00000007:error] !!! 1.5-7: syntax error: unexpected ;

We just upgraded bison so that now calls to yylex are wrapped with a
try/catch.  As a consequence we can now issue a genuine syntax error
for lexical errors:

  $ urbi -qe 'var 3x;shutdown;'
  [00000007:error] !!! 1.5-6: syntax error: invalid token: `3x'

	* src/parser/ugrammar.y: Require Bison 2.4 (actually we need
	2.4.600+).
	* src/parser/utoken.l (SCAN_ERROR): Here.
	(YY_FATAL_ERROR): Use it.
	(lexical_cast): Adjust.
	* doc/specs/lang.tex: Now we can truly exercise invalid tokens.