parser: don't treat invalid escapes as invalid tokens.
The commit:73c213daeb38e7a6b08ea6b8b859a288170a001c changed the handling
of "\h" as returning a parse error when meeting the \h escape.  This is
wrong on two accounts.  First, the scanner remains in string start-condition
so it looks "stuck".  Second, this is a regression for the user who
gets a single error at a time, where she used to be granted several at
once.

	* src/parser/utoken.l (SCAN_ERROR): Rename as...
	(SCAN_FATAL): this.
	(SCAN_ERROR): New.
	Use one or the other where appropriate.
	* tests: Straighten 2.x/string/escape.chk.
	* doc/specs/lang.tex, doc/specs/string.tex: Minor improvements.