Fix #4874: 1 in [Comparable] loops.
The problem is that Comparable is in the list of protos, and if you
try "foo in [Comparable]", you end up running "Comparable == foo".
And Comparable provides a default == which bounces to !=, and
conversely.  So, of course:

Comparable == 1;
[       Sched.Job       ] Stack space exhausted
[00105924:error] !!! 2.1-15: ==: stack space exhausted

Of course there is an issue here, but I'm not sure how to address this
in a nice way.  Yet, there's an easy way out: "x in [a, b, c...]"
actually check whether "a == x || b == x ||...".  It seems saner to
run "x == a || x == b || ..." etc., which solves the problem at hand.

	* share/urbi/list.u (has, hasSame): Keep the argument as first
	argument for comparisons.
	* doc/specs/list.tex: Fix wording.