doc: loops.
Enhance the documentation about the loops: more systematic description of
the difference flavors, improve the index about them, and provide a
table that points to the various sections.

Also, improve the outline for these constructs (i.e., more regular).
And simplify some tests to avoid scopes that should not bother our users.

	* doc/specs/lang.tex: Do that.
	Also, use more consistent labels for keywords.
	* doc/guide/faq.tex, doc/guide/k1-to-k2.tex,
	* doc/tables/release-notes.tex, doc/urbi-sandbox.tex:
	Adjust.
	* doc/document-aux: Update for index.
	* tests: merge tests about for.