ast: don't compute whether a file is a system file, know it.
We have failures on Windows like these:

-[01234567:error] !!! new: cannot convert to date: 2003-10-10  20:10:50
+[01234567:error] !!! init: cannot convert to date: 2003-10-10  20:10:50
+[01234567:error] !!!    called from: new

This is because the file is not recognized as a system file:

Date.new("2003-10-10  20:10:50");
[00006107:error] !!! c\object\object-class.cc:149.1: init: cannot convert to date: 2003-10-10  20:10:50
[00006107:error] !!!    called from: 1.1-32: new

The file name is broken because for some reason, now we get a
Windows-style value for __FILE__.  So instead of computing, from the hard
coded file name in system.u, whether the file is system, register system
files when their __FILE__ is needed.

	* src/object/system.cc (system_files_): To ensure proper
	initialization, make it a static variable from...
	(system_files_get): this.
	Adjust dependencies.
	* src/ast/loc.hh, src/ast/loc.cc (declare_location_file): New.
	(DECLARE_LOCATION_FILE): Use it.
	* share/urbi/system.u: Remove open-coded list of C++ system files.
	Simplify.
	* src/ast/local.mk: Adjust.

commit:434df9172d4814e5853b70d0b3d95dae9a3d68ba from 2.7.3.

Conflicts:

	share/urbi/system.u
	src/object/system.cc