Boost: require 1.40 instead of 1.38.
For some unexplained reasons, we had Windows builds in 2.6 that used Boost
1.39, whereas all the others were using 1.38 (including Ubuntu Lucid which
feature 1.40 as standard package...), while Urbi 2.7 on Windows used 1.38
(again?).

This is messy, and poorly documented.  But now that Naoqi uses a different
model to communicate with Urbi, we are no longer constrained by the
version of Boost we believe they still use (1.38).  So...

So, Matthieu and I believe that wherever Boost is available as a package,
use that package.  As a result, Lucid and Mac OS X move to 1.40. and
1.47. Debian Etch features 1.33 anyway, so instead of compiling 1.38
ourselves, let's compile 1.40 to match the Ubuntu.  Finally, on Windows,
the "packages" come from BoostPro, so install the latest: 1.47.

	* buildfarm/configure: Implement this.
	* doc/guide/build.tex: Doc it.
	* doc/tables/release-notes.tex: News it.