tests: skip tests that Windows fails.
Being able to run these tests is new, before we ran none of the tests that
used connections, since Wine did not support the feature Asio uses.  Now
Wine is good enough, so tests are activated.  Some do not pass, but it's
too late to investigate for 2.7.4.  Since there is no regression compared
to 2.7.3, proceed.

	* buildfarm/check, tests: here.