doc: ship tex2chk.
Issue #5111.

One reason for which "make" loops forever with in-place builds, is that
we don't ship tests/guide/requirements-test.mk, yet we have a rule to
build it, but it's not enabled because tex2chk is not available.  This
is wrong: users deserve to have all our tools.  But we also should
ship all the *mk files (another patch).

	* doc/Makefile.am, doc/tests/local.mk: Ship all our tools.